chore: move EnrichedTextInput spec to src/spec
